feat: Update Summary component to dynamically display total savings
The Summary component in the `(tabs)` directory has been updated to dynamically display the total savings based on the value of the `transactionState.totalBalance` property. If the total balance is greater than 0, the text color is set to "green"; otherwise, it is set to "red". This change improves the visual representation of the total savings in the Expense Tracker app.