refactor: Update Add.tsx component to support editing transactions
The Add.tsx component in the (tabs)/(transactions) directory has been updated to support editing transactions. This change includes adding the necessary logic to fetch and display the details of an existing transaction when the component is used for editing. It also includes updating the submit button text to "Update" when editing an existing transaction. This enhancement improves the functionality of the Expense Tracker app by allowing users to modify their transactions after they have been added.